What should have been in lab2.py in the first place
Instead of assigning a value to n as he was supposed to, in this file my partner simply re-wrote "Graded by" and his name. I had him give me a value of n, and edited it to set said value myself so I could actually use it in my quiz1.py.